requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0  # Faster than pandas for preprocessing
pyarrow>=13.0.0  # Parquet support

# Data Download
requests>=2.31.0
lxml>=4.9.3
beautifulsoup4>=4.12.0
yfinance>=0.2.28

# Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# RL (for Phase 2)
gymnasium>=0.29.0  # Replaces gym
stable-baselines3>=2.1.0

# Feature Engineering
ta-lib  # Technical Analysis Library (requires separate installation)
pandas-ta>=0.3.14  # Alternative if ta-lib fails

# Data Storage
h5py>=3.9.0
tables>=3.8.0  # PyTables for HDF5

# Hyperparameter Optimization
optuna>=3.3.0

# Experiment Tracking
mlflow>=2.7.0
tensorboard>=2.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0
joblib>=1.3.2

# Numerical Acceleration
numba>=0.58.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Production API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
gunicorn>=21.2.0

# MT5 Integration
MetaTrader5>=5.0.45

# Data Quality & Validation
scipy>=1.10.0

# Advanced Monitoring
wandb>=0.15.0

# Database
redis>=4.6.0
psycopg2-binary>=2.9.6

# Deployment
docker>=6.1.0
kubernetes>=27.2.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.8.5
